renamed screen template
